Fixed - default is not working for insert ordered list and insert unordered list #799
Issue: https://github.com/xdan/jodit/issues/799
