+++ test/test.html
@@ -45,7 +45,7 @@
 
 <script src="bootstrap.js"></script>
 
-<script src="tests/units/styleTest.js"></script>
+<script src="tests/units/style-test.js"></script>
 <script src="tests/units/popupTest.js"></script>
 <script src="tests/units/objectObserverTest.js"></script>
 <script src="tests/units/helpersTest.js"></script>
@@ -81,7 +81,7 @@
 <script src="tests/acceptance/plugins/wrap-nodes.js"></script>
 <script src="tests/acceptance/plugins/focus.js"></script>
 <script src="tests/acceptance/plugins/font.js"></script>
-<script src="tests/acceptance/plugins/orderedList.js"></script>
+<script src="tests/acceptance/plugins/ordered-list.js"></script>
 <script src="tests/acceptance/plugins/placeholder.js"></script>
 <script src="tests/acceptance/plugins/cleanHtml.js"></script>
 <script src="tests/acceptance/plugins/limit.js"></script>
